requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
mysqlclient==2.1.1
mysql-connector-python==8.0.27
python-dotenv==0.21.0
werkzeug==2.0.3
# functions==0.7.0
# pika==1.2.0
# requests==2.19.1
# SQLAlchemy==1.3.6
# twilio==7.7.1